videoplatform/pyproject.toml
[project]
name = "videoplatform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "autobahn==22.4.2",
    "autopep8>=2.3.2",
    "boto3>=1.42.73",
    "celery>=5.6.2",
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.2",
    "django>=6.0.3",
    "django-axes>=8.3.1",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.2.0",
    "django-environ>=0.13.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-imagekit>=6.1.0",
    "django-import-export>=4.4.0",
    "django-redis>=6.0.0",
    "django-rest-framework>=0.1.0",
    "django-storages>=1.14.6",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "factory-boy>=3.3.3",
    "graphene-django>=3.2.3",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "moviepy>=2.2.1",
    "nltk>=3.9.3",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "opencv-python-headless>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=11.3.0",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic>=2.12.5",
    "pymemcache>=4.0.0",
    "python-memcached>=1.62",
    "redis>=7.3.0",
    "stripe>=14.4.1",
]
